Add support for TFLite
I was trying to run some benchmarks and I was getting crashes with TFLite models. The important line is the isinstance check that can now receive a tf.lite.Interpreter object.

The rest is relatively straightforward since TFLite works directly on numpy arrays, but it is likely I am missing some extra things that need to be done to get this merged in.

I ran the TFLite models on the CPU and it doesn't seem that getting GPUs to work with TFLite would actually involve moving numpy arrays on/off device. I am also unsure what the GPU memory measurement would actually do in this case.

Let me know what you think!